model/requirement.txt
torch==1.7.0
pyro-ppl==1.4.0
lifelines==0.24.16
scikit-learn==0.22.2
pandas==1.0.5
seaborn==0.10.1
xlrd==1.2.0